+++ src/pretalx/common/forms/fields.py
@@ -1,13 +1,17 @@
+from io import BytesIO
 from pathlib import Path
 
 from django.conf import settings
 from django.contrib.auth.password_validation import validate_password
+from django.core.files import File
 from django.core.files.uploadedfile import UploadedFile
 from django.forms import CharField, FileField, ValidationError
 from django.utils.translation import gettext_lazy as _
+from PIL import Image
 
 from pretalx.common.forms.widgets import (
     ClearableBasenameFileInput,
+    ImageInput,
     PasswordConfirmationInput,
     PasswordStrengthInput,
 )
@@ -16,7 +20,9 @@ from pretalx.common.forms.widgets import (
 SVG image uploads were possible in the past, but have been removed due to
 security concerns.
 SVG validation is nontrivial to the point of not having been convincingly
-implemented so far, and always pose the risk of data leakage.
+implemented so far, and always pose the risk of data leakage. Additionally,
+Pillow does not support SVG images, and we use Pillow to strip metadata
+and reduce file size.
 
 As a compromise, users with the `is_administrator` flag are still allowed to
 upload SVGs, since they are presumed to have root access to the system.
@@ -48,7 +54,7 @@ class PasswordConfirmationField(CharField):
         super().__init__(*args, **kwargs)
 
 
-class SizeFileField(FileField):
+class SizeFileInput:
     """Takes the intended maximum upload size in bytes."""
 
     def __init__(self, *args, **kwargs):
@@ -66,22 +72,21 @@ class SizeFileField(FileField):
             num /= 1024
         return f"{num:.1f}YiB"  # Future proof 11/10
 
-    def clean(self, *args, **kwargs):
-        data = super().clean(*args, **kwargs)
+    def validate(self, value):
+        value = super().valiadate(value)
         if (
             self.max_size
-            and isinstance(data, UploadedFile)
-            and data.size > self.max_size
+            and isinstance(value, UploadedFile)
+            and value.size > self.max_size
         ):
             raise ValidationError(
                 _("Please do not upload files larger than {size}!").format(
                     size=SizeFileField._format_size(self.max_size)
                 )
             )
-        return data
 
 
-class ExtensionFileField(SizeFileField):
+class ExtensionFileInput:
     widget = ClearableBasenameFileInput
 
     def __init__(self, *args, **kwargs):
@@ -89,10 +94,9 @@ class ExtensionFileField(SizeFileField):
         self.extensions = [i.lower() for i in extensions]
         super().__init__(*args, **kwargs)
 
-    def clean(self, *args, **kwargs):
-        data = super().clean(*args, **kwargs)
-        if data:
-            filename = data.name
+    def validate(self, value):
+        if value:
+            filename = value.name
             extension = Path(filename).suffix.lower()
             if extension not in self.extensions:
                 raise ValidationError(
@@ -101,4 +105,76 @@ class ExtensionFileField(SizeFileField):
                     ).format(extension=extension)
                     + ", ".join(self.extensions)
                 )
-        return data
+        return value
+
+
+class SizeFileField(SizeFileInput, FileField):
+    pass
+
+
+class ExtensionFileField(ExtensionFileInput, SizeFileInput, FileField):
+    pass
+
+
+class ImageField(ExtensionFileInput, SizeFileInput, FileField):
+    # TODO: add better image field widget
+    widget = ImageInput
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, extensions=IMAGE_EXTENSIONS, **kwargs)
+
+    def to_python(self, data):
+        """
+        Check that the file-upload field data contains a valid image (GIF, JPG,
+        PNG, etc. -- whatever Pillow supports).
+
+        Vendored from django.forms.fields.ImageField to add EXIF data removal.
+        Can't use super() because we need to patch in the .png.fp object for
+        some unholy (and possibly buggy) reason.
+        """
+        f = super().to_python(data)
+        if f is None:
+            return None
+
+        # We need to get a file object for Pillow. We might have a path or we might
+        # have to read the data into memory.
+        if hasattr(data, "temporary_file_path"):
+            file = data.temporary_file_path()
+        else:
+            if hasattr(data, "read"):
+                file = BytesIO(data.read())
+            else:
+                file = BytesIO(data["content"])
+
+        try:
+            # load() could spot a truncated JPEG, but it loads the entire
+            # image in memory, which is a DoS vector. See #3848 and #18520.
+            image = Image.open(file)
+            # verify() must be called immediately after the constructor.
+            image.verify()
+
+            # Annotating so subclasses can reuse it for their own validation
+            f.image = image
+            # Pillow doesn't detect the MIME type of all formats. In those
+            # cases, content_type will be None.
+            f.content_type = Image.MIME.get(image.format)
+        except Exception as exc:
+            # Pillow doesn't recognize it as an image.
+            raise ValidationError(
+                self.error_messages["invalid_image"],
+                code="invalid_image",
+            ) from exc
+        if hasattr(f, "seek") and callable(f.seek):
+            f.seek(0)
+
+        stream = BytesIO()
+        stream.name = data.name
+        image.fp = file
+        if hasattr(image, "png"):  # Yeah, idk what's up with this
+            image.png.fp = file
+        image_data = image.getdata()
+        image_without_exif = Image.new(image.mode, image.size)
+        image_without_exif.putdata(image_data)
+        image_without_exif.save(stream)
+        stream.seek(0)
+        return File(stream, name=data.name)